+++ packages/eslint-config-airbnb-base/rules/es6.js
@@ -110,8 +110,7 @@ module.exports = {
 
     // Prefer destructuring from arrays and objects
     // http://eslint.org/docs/rules/prefer-destructuring
-    // TODO: enable
-    'prefer-destructuring': ['off', {
+    'prefer-destructuring': ['error', {
       VariableDeclarator: {
         array: false,
         object: true,