+++ packages/eslint-config-airbnb-base/rules/style.js
@@ -3,7 +3,7 @@ module.exports = {
     // enforce line breaks after opening and before closing array brackets
     // http://eslint.org/docs/rules/array-bracket-newline
     // TODO: enable? semver-major
-    'array-bracket-newline': ['off', { multiline: true, minItems: 3 }],
+    'array-bracket-newline': ['off', 'consistent'], // object option alternative: { multiline: true, minItems: 3 }
 
     // enforce line breaks between array elements
     // http://eslint.org/docs/rules/array-element-newline
@@ -36,6 +36,15 @@ module.exports = {
         ignoreInlineComments: true,
         ignoreConsecutiveComments: true,
       },
+    }],
+
+    // require trailing commas in multiline object literals
+    'comma-dangle': ['error', {
+      arrays: 'always-multiline',
+      objects: 'always-multiline',
+      imports: 'always-multiline',
+      exports: 'always-multiline',
+      functions: 'always-multiline',
     }],
 
     // enforce spacing before and after comma
@@ -73,6 +82,10 @@ module.exports = {
     // TODO: enable
     'func-style': ['off', 'expression'],
 
+    // enforce consistent line breaks inside function parentheses
+    // https://eslint.org/docs/rules/function-paren-newline
+    'function-paren-newline': ['error', 'multiline'],
+
     // Blacklist certain identifiers to prevent them being used
     // http://eslint.org/docs/rules/id-blacklist
     'id-blacklist': 'off',
@@ -91,9 +104,6 @@ module.exports = {
       VariableDeclarator: 1,
       outerIIFEBody: 1,
       // MemberExpression: null,
-      // CallExpression: {
-      //   parameters: null,
-      // },
       FunctionDeclaration: {
         parameters: 1,
         body: 1
@@ -101,7 +111,15 @@ module.exports = {
       FunctionExpression: {
         parameters: 1,
         body: 1
-      }
+      },
+      CallExpression: {
+        arguments: 1
+      },
+      ArrayExpression: 1,
+      ObjectExpression: 1,
+      ImportDeclaration: 1,
+      flatTernaryExpressions: false,
+      ignoredNodes: ['JSXElement', 'JSXElement *']
     }],
 
     // specify whether double or single quotes should be used in JSX attributes
@@ -134,6 +152,11 @@ module.exports = {
     // disallow mixed 'LF' and 'CRLF' as linebreaks
     // http://eslint.org/docs/rules/linebreak-style
     'linebreak-style': ['error', 'unix'],
+
+    // require or disallow an empty line between class members
+    // https://eslint.org/docs/rules/lines-between-class-members
+    // TODO: semver-major: enable
+    'lines-between-class-members': ['off', 'always', { exceptAfterSingleLine: false }],
 
     // enforces empty lines around comments
     'lines-around-comment': 'off',
@@ -178,6 +201,10 @@ module.exports = {
     // restrict the number of statements per line
     // http://eslint.org/docs/rules/max-statements-per-line
     'max-statements-per-line': ['off', { max: 1 }],
+
+    // enforce a particular style for multiline comments
+    // https://eslint.org/docs/rules/multiline-comment-style
+    'multiline-comment-style': ['off', 'starred-block'],
 
     // require multiline ternary
     // http://eslint.org/docs/rules/multiline-ternary
@@ -296,7 +323,7 @@ module.exports = {
     // disallow trailing whitespace at the end of lines
     'no-trailing-spaces': ['error', {
       skipBlankLines: false,
-      // ignoreComments: false, // TODO: uncomment once v3 is dropped
+      ignoreComments: false,
     }],
 
     // disallow dangling underscores in identifiers
@@ -304,7 +331,7 @@ module.exports = {
       allow: [],
       allowAfterThis: false,
       allowAfterSuper: false,
-      // enforceInMethodNames: false, // TODO: uncoment and enable, semver-minor once v3 is dropped
+      enforceInMethodNames: false,
     }],
 
     // disallow the use of Boolean literals in conditional expressions
@@ -325,10 +352,9 @@ module.exports = {
 
     // enforce line breaks between braces
     // http://eslint.org/docs/rules/object-curly-newline
-    // TODO: enable once https://github.com/eslint/eslint/issues/6488 is resolved and v3 is dropped
-    'object-curly-newline': ['off', {
-      ObjectExpression: { minProperties: 3, multiline: true, consistent: true },
-      ObjectPattern: { minProperties: 3, multiline: true, consistent: true }
+    'object-curly-newline': ['error', {
+      ObjectExpression: { minProperties: 4, multiline: true, consistent: true },
+      ObjectPattern: { minProperties: 4, multiline: true, consistent: true }
     }],
 
     // enforce "same line" or "multiple line" on object properties.
@@ -351,8 +377,8 @@ module.exports = {
     // enforce operators to be placed before or after line breaks
     'operator-linebreak': 'off',
 
-    // enforce padding within blocks
-    'padded-blocks': ['error', 'never'],
+    // disallow padding within blocks
+    'padded-blocks': ['error', { blocks: 'never', classes: 'never', switches: 'never' }],
 
     // Require or disallow padding lines between statements
     // http://eslint.org/docs/rules/padding-line-between-statements
@@ -377,8 +403,7 @@ module.exports = {
 
     // Enforce location of semicolons
     // http://eslint.org/docs/rules/semi-style
-    // TODO: enable, semver-major until v3 is dropped, semver-minor otherwise
-    'semi-style': ['off', 'last'],
+    'semi-style': ['error', 'last'],
 
     // requires object keys to be sorted
     'sort-keys': ['off', 'asc', { caseSensitive: false, natural: true }],
@@ -428,13 +453,11 @@ module.exports = {
 
     // Enforce spacing around colons of switch statements
     // http://eslint.org/docs/rules/switch-colon-spacing
-    // TODO: enable, semver-major
-    'switch-colon-spacing': ['off', { after: true, before: false }],
+    'switch-colon-spacing': ['error', { after: true, before: false }],
 
     // Require or disallow spacing between template tags and their literals
     // http://eslint.org/docs/rules/template-tag-spacing
-    // TODO: enable, semver-major
-    'template-tag-spacing': ['off', 'never'],
+    'template-tag-spacing': ['error', 'never'],
 
     // require or disallow the Unicode Byte Order Mark
     // http://eslint.org/docs/rules/unicode-bom