+++ packages/eslint-config-airbnb-base/rules/variables.js
@@ -16,8 +16,7 @@ module.exports = {
     'no-label-var': 'error',
 
     // disallow specific globals
-    // TODO: enable, semver-major
-    'no-restricted-globals': ['off'].concat(restrictedGlobals),
+    'no-restricted-globals': ['error', 'isFinite', 'isNaN'].concat(restrictedGlobals),
 
     // disallow declaration of variables already declared in the outer scope
     'no-shadow': 'error',